+++ R/ci.R
@@ -113,6 +113,9 @@ ci.draws <- function(x, ci = 0.95, method = "ETI", verbose = TRUE, BF = 1, ...) 
   .ci_bayesian(.posterior_draws_to_df(x), ci = ci, method = method, verbose = verbose, BF = BF, ...)
 }
 
+#' @export
+ci.rvar <- ci.draws
+
 
 #' @export
 ci.emmGrid <- function(x, ci = NULL, ...) {