+++ R/p_map.R
@@ -115,6 +115,9 @@ p_map.draws <- function(x, precision = 2^10, method = "kernel", ...) {
   p_map(.posterior_draws_to_df(x), precision = precision, method = method, ...)
 }
 
+#' @export
+p_map.rvar <- p_map.draws
+
 
 
 #' @export