+++ R/p_significance.R
@@ -107,6 +107,9 @@ p_significance.draws <- function(x, threshold = "default", ...) {
   p_significance(.posterior_draws_to_df(x), threshold = threshold, ...)
 }
 
+#' @export
+p_significance.rvar <- p_significance.draws
+
 
 #' @export
 p_significance.parameters_simulate_model <- function(x, threshold = "default", ...) {