+++ NEWS.md
@@ -4,8 +4,9 @@
 
 * `contr.equalprior(contrasts = FALSE)` (previously `contr.orthonorm`) no longer returns an identity matrix, but a shifted `diag(n) - 1/n`, for consistency.
 
-## New functions
+## New functionality
 
+* *bayestestR* now supports objects of class `rvar` from package *posterior*.
 * `contr.equalprior` (previously `contr.orthonorm`) gains two new functions: `contr.equalprior_pairs` and `contr.equalprior_deviations` to aide in setting more intuitive priors.
 
 ## Changes